+++ src/backend/app/users/__init__.py
@@ -1,1 +1,1 @@
-"""FMTM users API."""
+"""Field-TM users API."""