+++ src/frontend/src/utilities/CustomizedImage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import cardImg from '@/assets/images/project_icon.png';
-import logo from '@/assets/images/hotLog.png';
+import logo from '@/assets/images/favicon.svg';
 import { LazyLoadImage } from 'react-lazy-load-image-component';
 
 type switcherType = {